// _examples/golang-file-upload/interop/server.gen.ts
export interface FileServiceServer<Context = unknown> {
  uploadAvatar(ctx: Context, req: UploadAvatarRequest): Promise<UploadAvatarResponse>
  uploadAttachments(ctx: Context, req: UploadAttachmentsRequest): Promise<UploadAttachmentsResponse>
  downloadAvatar(ctx: Context, req: DownloadAvatarRequest): Promise<Blob | File>
  stampAvatar(ctx: Context, req: StampAvatarRequest): Promise<Blob | File>
  listAvatars(ctx: Context, req: ListAvatarsRequest): Promise<ListAvatarsResponse>
}

//
// Schema types
//

export interface UploadAvatarRequest {
  userId: number
  avatar: Blob | File
}

export interface UploadAvatarResponse {
  size: number
}

export interface UploadAttachmentsRequest {
  userId: number
  attachments: Array<Blob | File>
}

export interface UploadAttachmentsResponse {
  count: number
}

export interface DownloadAvatarRequest {
  userId: number
}

export interface StampAvatarRequest {
  userId: number
  stamp: Blob | File
}

export interface ListAvatarsRequest {
}

export interface ListAvatarsResponse {
  userIds: Array<number>
}
// ...
type WebrpcErrorParams = { name?: string, code?: number, message?: string, status?: number, cause?: string }

export class WebrpcError extends Error {
  code: number
  status: number

  constructor(error: WebrpcErrorParams = {}) {
    super(error.message)
    this.name = error.name || 'WebrpcEndpointError'
    this.code = typeof error.code === 'number' ? error.code : 0
    this.message = error.message || `endpoint error`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcError.prototype)
  }

  static new(payload: any): WebrpcError {
    return new this({ message: payload.message, code: payload.code, status: payload.status, cause: payload.cause })
  }
}


export class WebrpcEndpointError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'WebrpcEndpoint'
    this.code = typeof error.code === 'number' ? error.code : 0
    this.message = error.message || `endpoint error`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcEndpointError.prototype)
  }
}
// ...
export class WebrpcBadRouteError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'WebrpcBadRoute'
    this.code = typeof error.code === 'number' ? error.code : -2
    this.message = error.message || `bad route`
    this.status = typeof error.status === 'number' ? error.status : 404
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcBadRouteError.prototype)
  }
}
// ...
export class WebrpcBadRequestError extends WebrpcError {
  constructor(error: WebrpcErrorParams = {}) {
    super(error)
    this.name = error.name || 'WebrpcBadRequest'
    this.code = typeof error.code === 'number' ? error.code : -4
    this.message = error.message || `bad request`
    this.status = typeof error.status === 'number' ? error.status : 400
    if (error.cause !== undefined) this.cause = error.cause
    Object.setPrototypeOf(this, WebrpcBadRequestError.prototype)
  }
}